+++ autoload.php
@@ -8,7 +8,8 @@
 	*/
 	if(!class_exists("HttpRequest")) {
 		spl_autoload_register(function($class) {
-			$filename = __DIR__.DIRECTORY_SEPARATOR."src".DIRECTORY_SEPARATOR.$class.".class.php";
+			//$filename = __DIR__.DIRECTORY_SEPARATOR."src".DIRECTORY_SEPARATOR.$class.".class.php";
+			$filename = __DIR__.DIRECTORY_SEPARATOR.str_replace("\\", DIRECTORY_SEPARATOR, $class).".class.php";
 			if(is_readable($filename) && !class_exists($class)) {
 				require $filename;
 			}